fixing DBeaver compatibility

+++ src/main/java/com/example/bus_timetabling/repository/BusRepository.java
@@ -3,6 +3,6 @@ package com.example.bus_timetabling.repository;
 import com.example.bus_timetabling.entities.Bus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BusRepository extends JpaRepository<Bus, Integer> {
+public interface BusRepository extends JpaRepository<Bus, Long> {
 
 }
